send statics iformation to server and save static from server to model

+++ lib/models/home_model.dart
@@ -92,7 +92,7 @@ class User {
 
   final String fullName;
   final String avatarPath;
-  final int currentLevelProgress;
+  final double currentLevelProgress;
   final int level;
 
   factory User.fromJson(Map<String, dynamic> json) => User(
